Rename Text to TextWithLinks
